Revert "Merge branch 'SpeechRecognition' of https://github.com/180D-FW-2021/Team8 into SpeechRecognition"
This reverts commit bdd8670943e19fbc11d8b7fbda7e752d944eb0af, reversing
changes made to fe18954577ccc0ae668a277bd6e849c2ed5763a1.

+++ Assets/Scripts/PickMeScripts/ScoreCounter.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class ScoreCounter : MonoBehaviour
 {
-    public int score;
+    public float score;
     public float transitionSpeed = 100;
     float displayScore;
     public Text scoreText;
@@ -16,7 +16,7 @@ public class ScoreCounter : MonoBehaviour
         UpdateScoreDisplay();
     }
 
-    public void IncreaseScore(int amount)
+    public void IncreaseScore(float amount)
     {
         score += amount;
     }
